fix(bench): remove unnecessary casts in capture end extraction
Drop redundant `as usize` casts when reading capture end positions in benchmark loops.

This is a pure lint cleanup in benchmark code and does not affect behavior.

+++ benches/codspeed_bench.rs
@@ -588,7 +588,7 @@ fn bench_regression_scanner(c: &mut Criterion) {
                         ScannerFindOptions::NONE,
                     ) {
                         Some(m) => {
-                            let end = m.capture_indices[0].end as usize;
+                            let end = m.capture_indices[0].end;
                             pos = if end > pos { end } else { pos + 1 };
                             count += 1;
                         }
@@ -703,7 +703,7 @@ fn bench_regression_scanner_textmate(c: &mut Criterion) {
                         ScannerFindOptions::NONE,
                     ) {
                         Some(m) => {
-                            let end = m.capture_indices[0].end as usize;
+                            let end = m.capture_indices[0].end;
                             // Advance at least 1 position to avoid infinite loops
                             pos = if end > pos { end } else { pos + 1 };
                             count += 1;
@@ -868,7 +868,7 @@ fn bench_scanner_highlighting(c: &mut Criterion) {
                         ScannerFindOptions::NONE,
                     ) {
                         Some(m) => {
-                            let end = m.capture_indices[0].end as usize;
+                            let end = m.capture_indices[0].end;
                             pos = if end > pos { end } else { pos + 1 };
                             count += 1;
                         }
@@ -906,7 +906,7 @@ fn bench_scanner_highlighting(c: &mut Criterion) {
                         ScannerFindOptions::NONE,
                     ) {
                         Some(m) => {
-                            let end = m.capture_indices[0].end as usize;
+                            let end = m.capture_indices[0].end;
                             pos = if end > pos { end } else { pos + 1 };
                             count += 1;
                         }
@@ -960,7 +960,7 @@ fn bench_scanner_highlighting(c: &mut Criterion) {
                         ScannerFindOptions::NONE,
                     ) {
                         Some(m) => {
-                            let end = m.capture_indices[0].end as usize;
+                            let end = m.capture_indices[0].end;
                             pos = if end > pos { end } else { pos + 1 };
                             count += 1;
                         }
